Feature: Implement LoginPopup component for user authentication
Files:
- Added:  LoginPopup.js and loginPopup.css
- Updated: App.jsx, Navbar.js, main.jsx, nav.css and index.css

Description:
This feature introduces the LoginPopup component, enhancing the application's user authentication process. T

The LoginPopup.js file contains the implementation logic for handling user login actions and form submission. Styling for the popup window is defined in the loginPopup.css file.

In App.js, a state variable isLoginOpen is added to manage the visibility of the login popup. The toggleLogin function is implemented to toggle the state of the isLoginOpen variable and it passed as props to Navbar and Outlet components.

Navbar.js is updated to include a link for LoginPopup component .

In main.jsx, a Route is added to direct users to the RegisterPage component for user registration.

Additionally, styling for the login link in nav.css and related colors in index.css are updated.